convert ExplorerPage to TS
